add alert

+++ src/component/Restaurant.jsx
@@ -1,19 +1,38 @@
-import React from 'react'
-import Card from './Card';
-import axios from 'axios';
+import React from "react";
+import Card from "./Card";
+import axios from "axios";
+import Swal from "sweetalert2";
 
 const Restaurant = (props) => {
-
   const { restaurants, setEdit, setPopup } = props;
 
   console.log(restaurants);
 
-  
   const deleteRestaurant = async (id) => {
-    const response = await axios.delete(
-      `http://localhost:8080/restaurants/${id}`
-    );
-    return response.data;
+    const result = await Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    });
+
+    if (result.isConfirmed) {
+      Swal.fire({
+        title: "Deleted!",
+        text: "Your file has been deleted.",
+        icon: "success",
+      });
+
+      await new Promise((resolve) => setTimeout(resolve, 1000));
+
+      const response = await axios.delete(
+        `http://localhost:8080/restaurants/${id}`
+      );
+      return response.data;
+    }
   };
 
   return (
@@ -35,6 +54,6 @@ const Restaurant = (props) => {
       </div>
     </div>
   );
-}
+};
 
-export default Restaurant
+export default Restaurant;
